Refactoring: Move verification of citrus test in phase "verify".
We want to copy the logs from the citrus container before the maven execution fails. This makes it necessary to split the execution and verification of the citrus tests into different maven phases, so we can "move" the execution of docker:copy, docker:stop and docker:remove "between" test execution and test verification.