Refactoring: Give the docker:copy execution a separate stage.
In the current implementation, citrus-remote:test and docker:copy are executed within the same stage. This makes the order of execution dependent on the plugin order within the pom.xml. By providing a separate stage (post-integration-test), the only critical part is that docker:copy is executed before docker:stop and docker:remove. This makes the order of execution still dependent on the order in which the executions in the docker plugin are defined, but only within the same plugin.