#4: Fix progress bar and console reporting.
The RemoteTestResultReporter is unsuitable for a "live update" of test state. Switching to a test listener, however, brings the desired behaviour.

To get the missing time logging in the console logging, corresponding files and mappings have to be added.

There is still an open issue with the logging reporter; in the container, the PERFORMANCE shows "0ms", although each job shows its execution time.